backend-python/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# AI
groq==0.13.1

# Email
resend==2.22.0

# Real-time
python-socketio==5.12.1

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# HTTP client
httpx==0.28.1

# CORS & Security
python-multipart==0.0.20

# Dev
pytest==8.3.4
pytest-asyncio==0.25.0
